Добавил Interface
